requirements.txt
scipion-pyworkflow
scipion-em
scipion-em-motioncorr